requirements.txt
numpy>=1.21.0
scipy>=1.7.0
torch>=1.10.0
torchdiffeq>=0.2.2
matplotlib>=3.5.0
pandas>=1.3.0
PyYAML>=6.0
tqdm>=4.62.0
scikit-learn>=1.0.0
jupyter>=1.0.0
seaborn>=0.11.2
tensorboard>=2.7.0
hydra-core>=1.1.0 